Bug 1444192 - Prepare current content scripts for modularization.
Our current content scripts not only define functions, but they also run
code as soon as they get loaded. This was fine a while ago, but now we
have three independent scripts all doing things on their own, and
sometimes depend on global variables. To make sure we don't get
headache in the future, let's modularize those, and have a central
`run.js` that does the setting up and running.

This is just an intermediate step for moving files so we don't lose the
git history when changes are made to those files.