Bug 1444192 - Encapsulate modules and use dedicated run.js to run things.
This looks like a giant commit, but most of it is really just moving
code around so that all functions are enclosed in individual classes. I
also renamed the `background.js` to a more descriptive
`about_compat_broker.js`, because that's exactly what it does.

You will notice that we pass references to the interventions and user
agent override classes into the broker, and actually create a references
back to the broker. This isn't the best imaginable software design, but
it is an improvement to the dependency on global variables we previously
had. There are nicer methods to implement this, but since we are a bit
limited to what we can do inside webextensions and as we want to keep
the number of external dependencies as low as possible, that's probably
the best solution for now.