Bug 1444192 - Bump major version number.
Although this does not contain new features, I did touch basically all
the code imaginable, so if we'd have to backport things, it would be
very, very tricky. Therefore, let's bump the major version by one.

I also decided to set the version number in `package.json` to 1.0.0. We
frequently forget to touch that file, as the only important version
number is the one in `manifest.json`, so let's get rid of that second
step.