Issue #207 - drop web-ext.
We don't use that anymore, and the "Run the changed extension sources" documentation explicitly says to use `about:debugging`, which works. `web-ext` refuses to load the experimental APIs, even if I add `--pref=extensions.experiments.enabled=true` to the call.

The workflow with `about:debugging` works just as well, and you can even quickly build an .xpi to test, so let's just remove this dependency.

Closes #207.