Issue #256 - Add Jake task to automatically sync version numbers.
This task overwrites the version number in `package.json` with whatever is specified in `manifest.json`, to remove the need of manually syncing those up.

This task can be run via `npm run autoformat:version-number`, but is also run via the collection task `npm run autoformat`, which not only syncs the version number, but also ensures the correct code style, so that task is hopefully in everyone’s workflow already.