Issue #333 - Ensure all interventions have a globally unique ID.
Previously, the tests only verified that the ID is unique within each "kind" of intervention (i.e. UA override, Injection, Shim). This isn't reliable in practice.

The "disable" button on `about:compat` sends a message with `{ command: "toggle", id }` to the broker, and it doesn't distinguish between the different kinds of interventions. So the ID has to be globally unique, not just within each set.

This fixes the one instance where this happened, and also moves the test for this into a global test.

Fixes #333.